fiz algumas alterações e optimizações, o atv06 tava tudo podre

+++ cesar/numeros/Atv02.java
@@ -1,9 +1,9 @@
 // Crie um programa para encontrar a média dos números informados.
 public class Atv02{
     public static void main(String[] args) {
-        int total = 0;
+        float total = 0.0f;
         for(int i = 0; i < args.length; i++){
-            total += Integer.parseInt(args[i]);
+            total += Float.parseFloat(args[i]);
         }
         total /= args.length;
         System.out.println(total);
